Fix the gc problem.
If RegisterCoalescer join a copy and find the its allDefsAreDead,
it will try to eliminateDeadDefs of it. Then LRE will try to get its LIS
and splitSeparateComponents. So a joined register will be split by its
value defines. In this process, the recorded register is dead and have
no way to track down the value defines we are interested in.

So we have to create a new pass, run it after the RegisterCoalescer to
remove the registers from STATEPOINTs. We still benefit from the
removal. Or the Register Allocator may try to fulfill the constraint by
loading a value from stack.