Fix the gc problem.
If RegisterCoalescer join a copy and find the its allDefsAreDead,
it will try to eliminateDeadDefs of it. Then LRE will try to get its LIS
and splitSeparateComponents. So a joined register will be split by its
value defines. In this process, the recorded register is dead and have
no way to track down the value defines we are interested in.

We use the coloring and propagation algorithm, so that every value
should be colored, then the stack slot should be colored. Then we record
the stack slot colored. The phys registers do so too.