Fix CSR interference bug.
If a non-tagged live interval hits a call with an interference flag, we remove
the pointer registers from the usable registers set.
That's not enough for the case of a non-tagged live interval live before a
call with an interference flag. In such a case, the register will be
assigned and not reset. Then call to this interference function, it will
save this register as a CSR. So the marker will see this register's
value in the stack.
The new approach will clear the usable registers set of the registers in the
register mask of all interference calls in this function.