Fix unexpected taildup.
When a statepoint binds to a call site, and this callsite contains
object pool index, then a taildup dup this branch, the duplicated
statepoint uses the same index. It will cause mismatch when one of
the branch patch the object pool entry, and save its pc, and other
branch try to patch it, then pc mismatch.