Fix dart global variable addressing bug
LLVM assumes that the address of global variables are invariants.
However, dart is moving these global variables.
When lowering the node GlobalAddress, LLVM saves the result in the
NodeMap cache in the SelectionDAGBuilder and uses the result inside the block.
If the same global address is used in the same block multiple times and the use
points are spread across the call sites, the result of lowering will
be saved in a virtual register and used across these call sites.
So the result may be spilled in the stack slot and will be ignored by
the stack map generator.

The solution is to split the basic block at each call. We add a new pass
called `SplitStatepointBlock` to split immediate after the call to a
state point.